Add more game options

+++ krage.rb
@@ -1,3 +1,10 @@
+# frozen_string_literal: true
+unless RUBY_VERSION.to_f > 2.4
+  print `tput cup 22 70` unless `xprop -name Krage 2> /dev/null`.empty?
+  puts 'Require ruby > 2.4'
+  sleep 2 unless `xprop -name Krage 2> /dev/null`.empty?
+  exit
+end
 $LOAD_PATH.unshift(File.expand_path('lib', __dir__))
 require 'krage_class'
 
@@ -14,7 +21,6 @@ font = `dconf read /org/gnome/terminal/legacy/profiles:/:#{krage_profile_id}/fon
 curr_rows_cols = File.read("#{KRAGE_DIR}/ext/.current_rows_columns") rescue ''
 
 counter = 0
-
 if !`xprop -name Krage`.empty? && `stty size` != curr_rows_cols
   loop do
     rows, columns = rows_columns
@@ -39,8 +45,10 @@ music = ['krage_slavic.ogg', 'krage_western.ogg', 'krage_viking.ogg']
 spawn("paplay #{KRAGE_DIR}/data/#{music[rand(3)]}")
 
 at_exit do
-  `pkill -f 'paplay.*krage'`
-  `reset` if `xprop -name Krage`.empty?
+  if `xprop -name Krage`.empty?
+    `pkill -f 'paplay.*krage'`
+    `reset`
+  end
 end
 
 Signal.trap('INT') do
@@ -72,56 +80,54 @@ end
 def options_changer(options, left_right=false)
   options.each_with_index do |choice, idx|
     if choice[7] == '⮕'
-      pc_or_wifi = options[idx][-1].ascii_only? ? '' : options[idx][-1]
+      easy_hard = options[idx][-2].ascii_only? ? '' : options[idx][-2..-1]
       if left_right
-        pc_or_wifi = pc_or_wifi == '💻' ? '📶' : '💻'
-        options[idx][-1] = pc_or_wifi
+        easy_hard = easy_hard == '🔰 ' ? '⚔️' : '🔰 '
+        options[idx][-2..-1] = easy_hard
         break
       end
       other_idx = (idx-1).abs
-      reducer = pc_or_wifi.empty? ? -1 : -2
+      reducer = easy_hard.empty? ? -1 : -3
       options[idx] = options[idx][13..reducer]
-      options[other_idx] = POINTER + options[other_idx] + pc_or_wifi
+      options[other_idx] = POINTER + options[other_idx] + easy_hard
       break
     end
   end
 end
 
 def ending(player)
-  ending_choice = ["#{POINTER}Courageous enough for one more challenge",
-                   'I have had enough of this silly game']
-
   `pkill -f 'paplay.*krage'`
   player.spawn("paplay #{KRAGE_DIR}/data/king.ogg")
   loop do
     player.display('winer')
     puts "\n\n\n"
-    puts "\e[1;4;32mWe have a Krage King, behold almighty "\
+    puts "\e[1;4;34mWe have a Krage King, behold almighty "\
          "#{player.color.call(player.name.upcase)}!!!\e[0;1m\n\n".center(188)
 
-    options_writer(ending_choice)
+    options_writer(ENDING_CHOICE)
 
     key = `#{KRAGE_DIR}/ext/gen_keyboard`
     case key
     when ''
-      if ending_choice[0][7] == '⮕'
-        `pkill -f 'paplay.*krage'`
+      `pkill -f 'paplay.*krage'`
+      if ENDING_CHOICE[0][7] == '⮕'
         print "\ec"
         exec("ruby #{KRAGE_DIR}/krage.rb")
-      elsif ending_choice[1][7] == '⮕'
+      elsif ENDING_CHOICE[1][7] == '⮕'
         exit
       end
     when "\e[A", "\e[B"
-      options_changer(ending_choice)
+      options_changer(ENDING_CHOICE)
     end
   end
 end
 
-crow_file = File.read("#{KRAGE_DIR}/ext/crow.yaml")
-crow = YAML.load(crow_file)
+crow = YAML.load(File.read("#{KRAGE_DIR}/ext/crow.yaml"))
 
 POINTER = "\e[1;33m⮕\e[0m "
-game_options = ["#{POINTER}Krage For Softies 💻", 'Krage Under Pressure ']
+game_options = [+"#{POINTER}Krage For Softies 🔰 ", 'Krage Under Pressure ']
+ENDING_CHOICE = ["#{POINTER}Courageous enough for one more challenge",
+                 'I have had enough of this silly game']
 
 print "\e[?7l"
 intro(crow)
@@ -138,11 +144,8 @@ loop do
 
   case key
   when ''
-    if game_options[0][7] == '⮕'
-      Krage.class_variable_set(:@@game_with_timer, false)
-    elsif game_options[1][7] == '⮕'
-      Krage.class_variable_set(:@@game_with_timer, true)
-    end
+    Krage.class_variable_set(:@@game_hard, true) if game_options.any?(/⚔️/)
+    Krage.class_variable_set(:@@game_timer, true) if game_options[1][7] == '⮕'
     break
   when "\e[A", "\e[B"
     options_changer(game_options)
@@ -163,15 +166,17 @@ while num_players ||= `#{KRAGE_DIR}/ext/gen_keyboard`.to_i
       `stty echo`
       puts "\n\n"
       if num_players.between?(2, 4)
+        Krage::TWO_PL = num_players == 2 ? true : false
+        c, n = num_players < 3 ? [7, 4] : [2, 2]
         print "#{space}\e[1;40m1. Player Name\e[0;30m: "
         player1 = new_player
-        print "\n#{space}\e[1;47m2. Player Name\e[0;97m: "
+        print "\n#{space}\e[1;4#{c}m#{n}. Player Name\e[0;3#{c}m: "
         player2 = new_player
         if num_players > 2
-          print "\n#{space}\e[1;42m3. Player Name\e[0;32m: "
+          print "\n#{space}\e[1;41m3. Player Name\e[0;31m: "
           player3 = new_player
           if num_players == 4
-            print "\n#{space}\e[1;41m4. Player Name\e[0;31m: "
+            print "\n#{space}\e[1;47m4. Player Name\e[0;37m: "
             player4 = new_player
           end
         end
@@ -194,25 +199,26 @@ score = {}
 print "\e[?9h"
 print "\e[?7l"
 `stty -echo`
+save_state = `stty -g`
+`stty raw`
 print "\n\n" * (4 - players.size)
 print "\n#{space}\e[0mIacta alea est"
-5.times { print '!'; sleep 0.3 }
+5.times { print '!'; sleep 0.2 }
+`stty #{save_state}`
 
 players.cycle do |player|
   next unless player
   player.generate_coords == 'roll' ? player.roll : redo
-  score[player] = player.player_territory; next if player.round == 1
+  next score[player] = player.player_territory if player.round == 1
 
   loop do
     case button ||= player.generate_coords
     when true, 'skip'
-      player.show_current_land = false
-      player.countdown = '' if Krage.class_variable_get(:@@game_with_timer)
       player.spawn("paplay #{KRAGE_DIR}/data/skip.ogg") if button == 'skip'
       break
     when 'giveup'
-      player.show_current_land = false
       player.spawn("paplay #{KRAGE_DIR}/data/giveup.ogg")
+      score.delete(player)
       player.giveup_cleaner
       players[players.index(player)] = nil
       ending(players.compact.last) if players.count(&:itself) < 2
@@ -224,6 +230,9 @@ players.cycle do |player|
     end
     redo if button = player.choose_place
   end
+  player.show_current_land = false
+  player.countdown = '' if Krage.class_variable_get(:@@game_timer)
+  Krage.class_variable_set(:@@show_turn, true)
 
   players.compact.each do |pl|
     score[pl] = pl.player_territory
@@ -232,9 +241,9 @@ players.cycle do |player|
   next unless player == players.compact.last
   strongest = score.values.max
   second_strongest = score.values.max(2)[1]
-  empty_spaces = 900 - score.values.sum
+  empty_spaces = Krage.class_variable_get(:@@map).flatten.count('___')
 
-  if second_strongest + empty_spaces / (players.compact.size-0.75) < strongest
+  if second_strongest + empty_spaces / (players.compact.size-Krage::HANDICAP) < strongest
     ending(score.key(strongest))
   end
 end
